Rename files for Heroku deploy
