+++ src/__test__/components/Footer.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
-import Footer from '../../components/Footer';
+import Footer from '../../components/Footer/Footer';
 
 let getByTestId;
