+++ src/__test__/components/NotFound.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
-import NotFound from '../../components/NotFound';
+import NotFound from '../../components/NotFound/NotFound';
 
 let getByTestId;
 
@@ -10,7 +10,7 @@ beforeEach(() => {
   const component = render(
     <Router>
       <NotFound />
-    </Router>,
+    </Router>
   );
   getByTestId = component.getByTestId;
 });
@@ -19,7 +19,7 @@ it('renders the NotFound component', () => {
   const component = render(
     <Router>
       <NotFound />
-    </Router>,
+    </Router>
   );
 
   expect(component).toMatchSnapshot();
@@ -36,7 +36,7 @@ test('should render NotFound component with correct text', () => {
 
   expect(
     footerEl.textContent.includes(
-      'The page you are trying to view does not exits',
-    ),
+      'The page you are trying to view does not exits'
+    )
   ).toBe(true);
 });
