+++ src/components/CryptoCurrency/CryptoCurrency.js
