+++ src/containers/CryptoQuote/CryptoQuote.js
@@ -20,7 +20,7 @@ import { getCryptocurrencyQuoteAsync } from '../../redux/cryptoQuotesSlice';
 
 import styles from './CryptoQuote.module.css';
 import mainStyles from '../../index.module.css';
-import ErrorAlert from '../../components/Error';
+import ErrorAlert from '../../components/Error/Error';
 
 const CruyptoQuote = () => {
   const [cryptoQuote, setCryptoQuote] = useState(null);
