+++ src/index.js
@@ -4,7 +4,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import App from './containers/App';
+import App from './containers/App/App';
 import reportWebVitals from './reportWebVitals';
 import store from './redux/store';
